requirements.txt
absl-py==0.14.1
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.5
Django==3.2.6
django-cors-headers==3.8.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
geographiclib==1.52
geopy==2.2.0
haversine==2.4.0
idna==3.2
ortools==9.1.9490
Pillow==8.3.1
protobuf==3.18.0
pycodestyle==2.7.0
PyJWT==2.1.0
python-decouple==3.5
pytz==2021.1
razorpay==1.2.0
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6